requirements.txt
pymupdf
Pillow
deep-translator
